v-m-b #34 Fix Debian problems
Debian PIL returns dpi as non-ints ( (300.0, 300.0)
Debian JSON can't parse them
fix by casting to int before sending to json.dumps
