+++ .idea/inspectionProfiles/Project_Default.xml
@@ -32,6 +32,8 @@
           <option value="v_m_t.manifestforwork.arg_namespace" />
           <option value="localManifestForWork.VMBArgs.*" />
           <option value="manifestCommons.VMBArgs.*" />
+          <option value="PIL.Image.Image.open" />
+          <option value="manifestCommons.work_Rid" />
         </list>
       </option>
     </inspection_tool>
